feat: improve developer experience by sorting decades
The list of decades is now sorted chronologically when in development mode. This improves the developer experience by making the order of decades predictable.

refactor: update 2020s prompt to be more specific

The prompt for the 2020s has been updated to specify that the back of the mobile phone should be visible in the generated image. This will improve the quality of the generated images.